api/Python/requirements.txt
azure-common==1.1.28
azure-core==1.30.2
azure-cognitiveservices-speech==1.33.0
azure-cosmos==4.4.0b1
azure-identity==1.17.1
azure-search-documents==11.5.1
azure-storage-blob==12.19.0
backoff==2.2.1
beautifulsoup4==4.12.3
cssutils==2.9.0
fake-useragent==1.2.1
Faker==19.1.0
flask==3.0.0
langchain
langchain-core
langchain-experimental
langchain-community
langsmith==0.1.33
llama-index==0.11.1
llama-index-core==0.11.1
llama-index-llms-openai==0.2.0
llama-index-embeddings-openai==0.2.3
llama-index-embeddings-langchain==0.2.0
llama-index-embeddings-azure-openai==0.2.3
llama-index-llms-azure-openai==0.2.0
llama-index-vector-stores-azureaisearch==0.2.0
llama-index-agent-openai==0.3.0
llama-index-cli==0.3.0
llama-index-indices-managed-llama-cloud==0.3.0
llama-cloud==0.0.13
llama-index-multi-modal-llms-openai==0.2.0
llama-index-program-openai==0.2.0
llama-index-question-gen-openai==0.2.0
llama-index-readers-file==0.2.0
llama-index-readers-llama-parse==0.2.0
llama-parse==0.4.9
lxml==5.1.0
msal==1.30.0
msal-extensions==1.2.0
nltk==3.9.1
numpy==1.24.2
openai
openpyxl==3.1.2
pandas==1.5.3
pathos==0.2.9
pdfminer.six==20221105
Pillow==10.1.0
pytz==2022.7.1
pypdf==4.0.1
ratelimit==2.2.1
regex==2023.6.3
requests==2.31.0
requests-oauthlib==1.3.1
tenacity==8.2.2
tiktoken==0.5.2
tqdm==4.66.1
urllib3==2.0.4
langchain-openai==0.1.1
langchainhub==0.1.15
langchain-pinecone==0.0.3
